Fixes #4. The syntax for msgs with many arguments is <msg_name> \< <type1> <arg1> <type2> <arg2> <etc> \>
It is necessary to give the type of the arguments to
construct the Send procedure.

This commit includes also a first version of the back-end AST.
It relies heavily on the murphi skeleton in the MSI protocols
implementations. It is almost complete; the datatypes for the Starting
State and the Invariants are missing.

Lastly there are some minor improvements to the data types for the
grammar of the front-end