Fixes #24
When we want to define a non-symmetric machine we just use the word
"nonsymmetric" before the declaration e.g.
machine nonsymmetric Directory[2] { <etc.> }
If a machine has size 1, we will access directly the single element of
the machine, as we would do for non-symmetric ones with bigger size.
That's why we parse machines of size 1 as non-symmetric. The syntax is
simpler: we just write the name of the machine, without the size in
brackets or "nonssymmetic". For example:
machine Directory { <etc.> }