File restructuring
Now there is one file for each of the 8 parts of the target murphi
program. Each has the tomurphi implementation of the relevant part.
Previously all tomurphi implementations where in MurphiPrint, which has
now only tomurphi for the top program, which calls all other
implementations.
There are two more files with helper functions, one with tomurphi
implementations and another with general helper functions.
Furthermore, there have been significant changes to the Types section,
which is now in accordance with the new types for non-symmetric
machines.