Debugged Types
Types, that was previously changed to be in accordance with the new
datatype for non-symmetric machines, had bugs and some design issues.

In MurphiAST, the Types datatype has been altered and a new Symmetry
enum is used to signal whether a machine is symmetric or not.

In GenHelper fst3, snd3 and thrd3 were implemented for 4-tuples instead
of 3. This has been corrected.