Fixes #25
The backend now fully supports non-symmetric machines.
Most of the work was done in the last commit. In retrospect, the only
parts of the code that were affected were the MurphiAST and the helper
functions, which take care of the "low-level" printing.
The changes here are minor and mostly due to renaming helper functions.