Nonsymmetric machines index
The index of nonsymmetric machines has type
NonsymMIndex = Arbitrary
	     | Specific Index
Arbitrary means we refer to an arbitrary nonsymmetric machine inside a
function definition, a ruleset etc. where there is a formal parameter
declared for the machine index
Specific Index is used when we refer explicitly to a specific instance
of the machine (after all that's why we need nonsymmetric machines
anyway). For example, we may want to send a message to a specific
directory.