Fixes #31. Updated syntax for msg arguments
Now the syntax for message arguments is as discussed in the comments of
issue 31, with a small deviation. The message arguments have the form
<type decl> = <var name>
or just
<type decl>
In the  <type decl> we declare the parameter of the message i.e. its
type and name.
The <var name> is the variable passed as argument.
In the common case when the variable passed is the same as the parameter
name, we use the second form.