This fixes #33, fixes #34 : declaring channels, single nets, send/receive and channels
We don't need anymore to declare channels, because they are implicitly
declared with their respective networks.

When there is a single network in the model, we don't need to give a
name for it. We just write whether it is ordered or unordered and it
channels i.e. we just omit its name.

When we send/receive a message we add a declaration for the channel the
message is travelling through. The syntax is
\@ <channel name>
and we add it after the send/receive e.g.
src!Ack@chan1