Fixes #38. Message args as guards
The way the datatype MsgArg works it is more convenient not to have
separate datatypes for when an equals is used as an assignment or a
test. We can tell which way it is used when we transform the AST.
This commit has slightly different constructors for MsgArg to reflect
this. The problem will be really solved once we update the equivalent
datatype in the backend and we complete the relevant step of the
translation.