Fixes #36, fixes #37. VC is arg of Send, Receive and guards can contain arbitrary msg args
The channel through which we send a message is now an argument of the
constructors Send and Receive. In the latter it is wrapped in a Maybe,
because we may not need it in a guard.

Receive has also a list of tuples with msg args and the values
(constants or fields) they must have i.e. they function as conditions.
So now we can express arbitrary conditions involving message arguments
and the channel a message is sent through.

The relevant printing function have also been updated.