Fixes #46. Implemented GetTypes
This commit consists of the implementation of functions for building the
Types section in MurphiAST. Some of the functions are in TransGen.hs as
they extract general information from the front-end AST (e.g. fields of
machines) and thus may be used in other files.
MurphiAST has very minor modifications.