Fixes #47. Changes in CommonFunctions.hs
The CommonFunctions part of the MurphiAST has a field for the vc through
for broadcasting messages. We didn't introduce a datatype but we have a
triplet instead of a pair to capture the necessary information, as this
extension was easier. Also a small bug was fixed in the printing of
broadcasting functions.