Closes #42, closes #53. Parsing array elements, finding local vars
TransGen has a function for finding the local variables in a list of
responses. From this it is very easy to extend it to rules and machine
functions.

Expressions of the form \<identifier\> \[ \<num\> \] are parsed as array
elements by default. The translator changes them to instances of
machines by going through the list of machines. If the identifier
matches the name of a symmetric machine an error is thrown, because we
can't access directly symmetric machines.