Closes #51. Function from F.Param to B.Field, minor changes in Ast and MurphiAst
In TransGen we implemented a functions that, given a variable in the
frontend it derives its owner and transforms it to the appropriate
constuct in the backend.
The type Machine in MurphiAST has a new constructor
AnyType \<machinetype\> in the cases we don't need to know whether the
machine is symmetric or not.
There is also some refactoring in Ast.