Closes #51. Transforming messages
We implemented a function that given a message in the frontend and other
information, necessary to infer the owners of variables passed as
message arguments, returns a message in the backend. This information is
the machine in which the message is used, the arguments of the machine,
the local variables for this section and the general form of a backend
message i.e. a list of B.MsgArg that has all the arguments of backend
messages.