Implemented TransResponse
In the file TransResponse we implement a method for translating
responses (i.e. the actions that a machine does as a response to
receiving a message). This is the bulk of coherence protocols.

In Ast.hs and Parser.y we changed a small detail: the identifiers for
integer variables are parsed with type Param. This helps with the
translation and is also more scalable.