Closes #58. Machine indices vc indexed machine arrays
We introduce a constructor for Field that is printed to a machine index
(taking care for nonsymmetic ones) and introduced a function in
GenHelper.hs for transforming arbitrary instances of Field to this
constructor, if they refer to a machine. That is because at the
translation level, to keep things simple we translate frontend
constructs only to the first general constructor of Field (also called
Field). Then at the tomurphi implementations of Send, Broadcast, Add,
Remove (the last two are for sets) we locally use this function to
ensure we have an index and not an array if we refer to machines
directly and not with variables.
The functions for advancing queues also need indices but these are
printed directly and they don't need a datatype.