Closes #60. EmptyResp in the Backend, no other compile  bugs
In the response datatype we added another constructor EmptyResp, with
argument a string. It matches to the frontedn AST  constuctor EmptyResp,
and their purpose is to enclose strings important for understanding a
protocol that carry no meaning in murphi like e.g. "hit".

After this bug fix the compilation happens with no problems.