Changed the name of machine states to cover all cases.
The state of a machine "foo" is "foo"_<front state> where <front state>
is the name of this state in the frontend. This way we can be sure no
problems arise in murphi by machines that have states with the same name
in the frontend, even if the machines start with the same letter.