Closes #68. Arrays are transformed to machine instances as required
The approach we followed was to include one more argument to the
function trasVar, defined in TransGen, which transforms a variable in
the frontend to a parameter in the backend (they are the constructs that
hold values and are passed as arguments to messages or to Send). The way
transVar works is keep the context information we need to print a
variable, because in murphi we need to know who owns this variable. This
information includes now the list of the nonsymmetric machines, so that
we can transform array elements to instances of nonsymmetric machines if
needed. The reason this may be the case is because they have the same
syntax and the parser parses them by default as array elements and lets
the translator decide what they are.