Broadcast has the same syntax with Send. Closes #80.
Now, if we want to broadcast a message to a set of machines we simply
use the same syntax as for send : dst!msg@channel.

The translator checks if the dst is a set. If it is, the Send is
translated to a broadcast. We don't check at the translator that the set
contains appropriate machines, since this is caught by murphi.