Prints prettier murphi code. Check msg ars, add to set fix
Changed Cast to Broadcast

Minor printing change in Startstate.hs

When we check for message arguments in Machine functions, we have
checked the src before. This is to avoid murphi complaining about
undefined fields. This Fixes #91

The function for adding an element to a set worked incorrectly. It added
only elements that are already in the multiset instead of elements that
are not in it.