One broadcast function for each set, msg and machine index only arguments.
Now there is one broadcast function for each set we broadcast to at the
input. The message we broadcast is an argument to the function. Contrast
this with the previous situation: there was a broadcasting function for
each pair of message and dst set. What allows us to pass the message as
argument is the form of Send, as it takes a msg as argument.

Fixes #92