Numbers cannot be parsed as nums now
Previously, the derivation rule for param allowed a num to be parsed as
both a param or an int exp.

The only downside of the change is that if we want to assign a numeric
value to a message argument we must do so with a local variable. Perhaps
in a future version we can allow assignments inside message arguments to
be int exps. This would require some changes to the front ast and
possibly also to the backend and the translation procedure.