Target file name can include ".m"
Previously the name of the target file produced from the translator
should be specified without the extension. The translator added
automatically ".m" in the end. Now we check for the occurrence of an
ending in the target name given as argument. If there is one, we keep
the name given by the user otherwise we add ".m".